requirements.txt
av==11.0.0
einops
flashy>=0.0.1
hydra-core>=1.1
hydra_colorlog
julius
num2words
numpy<2.0.0
sentencepiece
spacy==3.7.6
huggingface_hub
tqdm
transformers>=4.31.0
xformers<0.0.23
demucs
librosa
soundfile
gradio
torchmetrics
encodec
protobuf
pesq
pystoi
torchdiffeq